+++ packages/storage/src/keyset.rs
@@ -417,7 +417,6 @@ impl<'a, K: Serialize + DeserializeOwned, Ser: Serde> Keyset<'a, K, Ser, WithIte
     /// paginates only the values.
     pub fn paging(&self, storage: &dyn Storage, start_page: u32, size: u32) -> StdResult<Vec<K>> {
         let start_pos = start_page * size;
-        let mut end_pos = start_pos + size - 1;
 
         let max_size = self.get_len(storage)?;
 
@@ -429,43 +428,12 @@ impl<'a, K: Serialize + DeserializeOwned, Ser: Serde> Keyset<'a, K, Ser, WithIte
             return Err(StdError::NotFound {
                 kind: "out of bounds".to_string(),
             });
-        } else if end_pos > max_size {
-            end_pos = max_size - 1;
         }
-        self.get_keys_at_positions(storage, start_pos, end_pos)
-    }
 
-    /// tries to list keys without checking start/end bounds
-    fn get_keys_at_positions(
-        &self,
-        storage: &dyn Storage,
-        start: u32,
-        end: u32,
-    ) -> StdResult<Vec<K>> {
-        let start_page = self.page_from_position(start);
-        let end_page = self.page_from_position(end);
-
-        let mut res = vec![];
-
-        for page in start_page..=end_page {
-            let indexes = self.get_indexes(storage, page)?;
-            let start_page_pos = if page == start_page {
-                start % self.page_size
-            } else {
-                0
-            };
-            let end_page_pos = if page == end_page {
-                end % self.page_size
-            } else {
-                self.page_size - 1
-            };
-            for i in start_page_pos..=end_page_pos {
-                let key_vec = &indexes[i as usize];
-                let key = self.deserialize_key(key_vec)?;
-                res.push(key);
-            }
-        }
-        Ok(res)
+        self.iter(storage)?
+            .skip(start_pos as usize)
+            .take(size as usize)
+            .collect()
     }
 
     /// Returns a readonly iterator only for values.
@@ -1058,6 +1026,24 @@ mod tests {
         assert_eq!(iter.next(), Some(Ok(4321)));
         assert_eq!(iter.next(), Some(Ok(3412)));
         assert_eq!(iter.next(), None);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_keymap_paging_last_page() -> StdResult<()> {
+        let mut storage = MockStorage::new();
+
+        let total_items: u32 = 20;
+        let keyset: Keyset<u32> = Keyset::new(b"test");
+
+        for i in 0..total_items {
+            keyset.insert(&mut storage, &i)?;
+        }
+
+        assert_eq!(keyset.paging(&storage, 0, 23)?.len(), 20);
+        assert_eq!(keyset.paging(&storage, 2, 8)?.len(), 4);
+        assert_eq!(keyset.paging(&storage, 2, 7)?.len(), 6);
 
         Ok(())
     }